requirements.txt
opencv-python>=4.9
numpy>=1.26
python-dotenv>=1.0
